rpg/src/main/java/dms/pastor/prototypes/dcs/spells/BubbleShieldSpell.java
package dms.pastor.prototypes.dcs.spells;

import dms.pastor.prototypes.dcs.Elements;
import dms.pastor.prototypes.dcs.conditions.ConditionEntry;
import dms.pastor.prototypes.dcs.conditions.ConditionType;
import dms.pastor.prototypes.dcs.units.Unit;

/**
 * Author Dominik Symonowicz
 * Created 04/08/2017
 * WWW:	https://dominiksymonowicz.com/welcome
 * IT BLOG:	https://dominiksymonowicz.blogspot.co.uk
 * GitHub:	https://github.com/pastorcmentarny
 * Google Play:	https://play.google.com/store/apps/developer?id=Dominik+Symonowicz
 * LinkedIn: https://www.linkedin.com/in/dominik-symonowicz
 */
public class BubbleShieldSpell extends Spell {

    private static final int INITIAL_TURNS_LEFT = 20;

    public BubbleShieldSpell() {
        setName("Bubble shield");
        setElements(new Elements(3, 1, 1, 1));
    }

    @Override
    public void castSpell(Unit attacker, Unit defender) {
        castSpellMessage(attacker.getName(), getName(), defender.getName());
        attacker.getConditions().add(ConditionEntry.createTemporaryCondition(ConditionType.BUBBLE_SHIELD, INITIAL_TURNS_LEFT));
    }
}
